% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranking_functions.R
\name{avg_rank}
\alias{avg_rank}
\title{Average of rank values across multiple columns.}
\usage{
avg_rank(reserves, columns_asc = c(), columns_desc = c())
}
\arguments{
\item{reserves}{table used for ranking.}

\item{columns_asc}{vector of column names that should be ranked in ascending order}

\item{columns_desc}{vector of column names that should be ranked in descending order}
}
\value{
A vector of values containing the summed ranks across all requested columns.
}
\description{
For a given table and a list of columns, calculates the rank of each reserve in each column, and returns the average.
}
\details{
NA values are ranked last. Ties receive the same value, calculated as the average of their collective indices.
}
\examples{
tbl <- data.frame(col1 = c(1,2,3,4,5), col2 = c(0.4, 0.5, 0.2, 0.05, 0.9), col3 = c(10, NA, 8, 11, NA))
avg_rank(tbl, columns_asc = c('col1', 'col3'), columns_desc = 'col2')
}
