% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{benchmark_table_sample}
\alias{benchmark_table_sample}
\title{Sample benchmarks table.}
\format{
A data.frame object with 3 columns and 24 rows.
}
\source{
\url{https://beaconsproject.ualberta.ca/}
}
\usage{
benchmark_table_sample
}
\description{
A tibble containing the name of three fictional benchmarks as the column names (PB_0001, PB_0002, PB_0003),
and values representing the catchment CATCHNUMs making up each benchmark.
As per the BUILDER output table, NA is used to make even column lengths.
}
\keyword{datasets}
