% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dissimilarity.R
\name{calc_dissimilarity}
\alias{calc_dissimilarity}
\title{Calculate dissimilarity values for a set of polygons.}
\usage{
calc_dissimilarity(
  reserves_sf,
  reference_sf,
  raster_layer,
  raster_type,
  categorical_class_values = c(),
  plot_out_dir = "",
  categorical_class_labels = data.frame()
)
}
\arguments{
\item{reserves_sf}{sf object with unique id column named \code{network}}

\item{reference_sf}{sf object of the reference area to compare against.}

\item{raster_layer}{Raster object that will be clipped to the reference and reserve areas, with crs matching reserves_sf}

\item{raster_type}{'categorical' will use Bray-Curtis, 'continuous' will use KS-statistic.}

\item{categorical_class_values}{Vector of raster values in \code{raster_layer} of type 'categorical' to include in the calculation.
Allows unwanted values to be dropped. Defaults to include all non-NA values.}

\item{plot_out_dir}{Path to folder in which to save plots. Default is not to create plots.
Only creates plots if valid file path is provided. Dir will be created if it doesn't exist.}

\item{categorical_class_labels}{Optional data.frame object with columns \code{values} and \code{labels} indicating the label to use in Bray-Curtis graphs
for each raster value. Defaults to using the raster values. Labels can be provided for all or a subset of values. See examples.}
}
\value{
A vector of dissimilarity values matching the order of the input \code{reserves_sf}. Optionally a dissimilarity plot saved in the
\code{plot_out_dir} for each computed value.
}
\description{
For a list of features (e.g. benchmarks or networks), calculate the dissimilarity value between each feature and a reference area for the
provided raster layer. Continuous rasters use the KS-statistic to compare distributions, categorical rasters use the Bray-Curtis
statistic.
}
\details{
Graphs comparing distributions can optionally be created and saved to a user provided file path.

NA values are always removed. For categorical rasters, values can optionally be subset for the calculation and graphs using
\code{categorical_class_values}.
}
\examples{
reserves <- catchments_to_benchmarks(
  benchmark_table_sample, 
  catchments_sample, 
  c("PB_0001", "PB_0002", "PB_0003"))
calc_dissimilarity(reserves, ref_poly, led_sample, 'categorical')
calc_dissimilarity(reserves, ref_poly, led_sample, 'categorical', c(1,2,3,4,5))
calc_dissimilarity(reserves, ref_poly, led_sample, 'categorical', c(1,2,3,4,5), 
  "temp/plots", data.frame(values=c(1,2,3,4,5), labels=c("one","two","three","four","five")))
}
