% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dci.R
\name{calc_lwdci}
\alias{calc_lwdci}
\title{Calculate stream network length-weighted Dendritic Connectivity Index for a set of reserves.}
\usage{
calc_lwdci(reserve_sf, stream_sf, buffer_width = 0.1)
}
\arguments{
\item{reserve_sf}{sf object of reserves in which to calculate DCI.}

\item{stream_sf}{sf object of river network. Must have streams grouped in a BASIN attribute.}

\item{buffer_width}{Width of buffer to apply to stream segments. Defaults to 0.1. Used to
ensure adjacent stream segments are connected during analysis.}
}
\value{
Vector of numeric DCI values matching the input features.
}
\description{
Dendritic connectivity measures the longitudinal connectivity of a river network inside a set of reserves
based on the expected probability of an organism being able to move between two random points in the
network \url{https://link.springer.com/article/10.1007/s10980-008-9283-y}{Cote et al 2009}.
}
\details{
DCI = sum ( li2 / L2 )
li - length of stream section
L - total length of all stream sections in benchmark

Values range between 0 and 1, with 1 indicating a completely connected river network.

For length weighted DCI, a separate DCI measure is calculated for each group of BASIN streams
within the reserve, then a weighted average is calculated where the weights are the lengths
of streams in each BASIN.
}
\examples{
reserves <- catchments_to_benchmarks(
  benchmark_table_sample, 
  catchments_sample, 
  c("PB_0001", "PB_0002"))
calc_lwdci(reserves, streams_sample)

}
