% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{catchments_sample}
\alias{catchments_sample}
\title{Sample of catchments with LED areas in km2.}
\format{
A sf object with 87 features and 16 fields.
}
\source{
\url{https://beaconsproject.ualberta.ca/}
}
\usage{
catchments_sample
}
\description{
A subset of the BEACONs catchments dataset intersecting the led_sample raster.
LED areas (in km2) have been added to the sample catchments using criteria_to_catchments().
}
\keyword{datasets}
