% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{catchments_to_networks}
\alias{catchments_to_networks}
\title{Dissolve a set of catchments to create benchmark polygons.}
\usage{
catchments_to_networks(benchmark_table, catchments_sf, network_list)
}
\arguments{
\item{benchmark_table}{Data frame where columns are benchmark names and rows are catchments making up the benchmark. e.g. the "COLUMN_All_Unique_BAs" table
output by BUILDER.}

\item{catchments_sf}{sf object of the catchments dataset with unique identifier column: CATCHNUM}

\item{network_list}{Vector of networks to build - usually just the list of benchmarks in the benchmark_table (e.g. PB_0001), but can also be combinations of
benchmarks (e.g. PB_0001__PB_0002) made using \code{gen_network_names()} (usually it is more efficient to build individual benchmarks then merge them into
networks using \code{benchmarks_to_networks()}).}
}
\value{
A sf object of networks
}
\description{
Takes a list of catchments defining a benchmark or a network of multiple benchmarks, and dissolves the
catchments into a single polygon feature.
}
\examples{
library(sf)

# Individual benchmarks
benchmarks <- catchments_to_networks(
  benchmark_table_sample, 
  catchments_sample, 
  c("PB_0001", "PB_0002"))
plot(benchmarks)

# Benchmarks combined into a network
network <- catchments_to_networks(
  benchmark_table_sample, 
  catchments_sample, 
  c("PB_0001__PB_0002"))
plot(network)
}
