% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criteria_to_catchments.R
\name{criteria_to_catchments}
\alias{criteria_to_catchments}
\title{Sum areas of criteria raster values inside catchment polygons.}
\usage{
criteria_to_catchments(
  catchments_sf,
  criteria_raster,
  criteria_name,
  class_vals = c()
)
}
\arguments{
\item{catchments_sf}{sf object of catchments}

\item{criteria_raster}{Raster object of the criteria layer that will be summed, with crs matching catchments}

\item{criteria_name}{String representing the criteria name that will provide the suffix in the new column names}

\item{class_vals}{Vector of the class values to sum. Defaults to including all unique values of the raster intersecting the catchments}
}
\value{
sf object matching catchments_sf, with the additional columns added and areas reported in km2
}
\description{
For a given raster layer, sums the area of all unique values in the raster and adds summed area's as columns in the catchments dataset.
}
\details{
Raster projection units are assumed to be in metres.
}
\examples{
criteria_to_catchments(catchments_sample, led_sample, "led")
}
