% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrology.R
\name{dissolve_upstream_catchments}
\alias{dissolve_upstream_catchments}
\title{Takes a table listing the upstream catchments for a set of reserves, dissolves the catchments and calculates
upstream area. Optionally calculates area-weighted intactness.}
\usage{
dissolve_upstream_catchments(
  catchments_sf,
  upstream_table,
  out_feature_id,
  intactness_id = ""
)
}
\arguments{
\item{catchments_sf}{sf object of the catchments dataset with unique identifier column: CATCHNUM .}

\item{upstream_table}{Data frame where column names are unique reserve names and rows are catchments making up the upstream area.
For example, the UPSTREAM_CATCHMENTS_COLUMN table from BUILDER, or the output of \code{\link[=get_upstream_catchments]{get_upstream_catchments()}}.}

\item{out_feature_id}{Name of the unique identifier column in output (e.g. "networks").}

\item{intactness_id}{Optional string identifying an intacntess column (values between 0 and 1) in catchments_sf. If provided,
used to calculate the area weighted intactness (AWI) of the upstream polygon.}
}
\value{
An sf object of upstream areas for the input features with \code{area_km2} and (optionally) \code{AWI} values added. Only
features from \code{upstream_table} that have valid catchments in \code{catchments_sf} will be included.
}
\description{
Takes a table listing the upstream catchments for a set of reserves, dissolves the catchments and calculates
upstream area. Optionally calculates area-weighted intactness.
}
\examples{
reserves <- catchments_to_benchmarks(
  benchmark_table_sample, 
  catchments_sample, 
  c("PB_0001", "PB_0002", "PB_0003"))
upstream_table <- get_upstream_catchments(reserves, "network", catchments_sample)
dissolve_upstream_catchments(catchments_sample, upstream_table, "network")
}
