% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/representation.R
\name{evaluate_targets_using_benchmarks}
\alias{evaluate_targets_using_benchmarks}
\title{Combines benchmark results to evaluate networks.}
\usage{
evaluate_targets_using_benchmarks(benchmark_results, network_list)
}
\arguments{
\item{benchmark_results}{Tibble or data frame output from \code{\link[=evaluate_targets_using_catchments]{evaluate_targets_using_catchments()}} or \code{\link[=evaluate_targets_using_clip]{evaluate_targets_using_clip()}}}

\item{network_list}{Vector of networks to evaluate, such as those made with \code{\link[=gen_network_names]{gen_network_names()}}.}
}
\value{
A tibble with columns:
\itemize{
\item{\code{class_value}: the list of class_values from the target_table}
\item{\code{area_km2}: area of the class value in the network}
\item{\code{network}: name of the network copied from network_list}
\item{\code{class_proportion}: copied from target_table}
\item{\code{target_km2}: copied from target_table}
\item{\code{prop_target_met}: area_km2/target_km2}
}
}
\description{
For a set of benchmarks in a network, takes the largest area value for each target class and evaluates the target. If area is
tied, selects one at random.
}
\details{
Each target therefore has to be met in full by at least one benchmark in the network. To assess target classes summed
across the entire network, use \code{\link[=evaluate_targets_using_catchments]{evaluate_targets_using_catchments()}} or \code{\link[=evaluate_targets_using_clip]{evaluate_targets_using_clip()}}.
}
\examples{
target_table <- gen_targets(ref_poly, led_sample, 1600)
benchmarks_results <- evaluate_targets_using_catchments(
  catchments_sample, 
  "led", 
  benchmark_table_sample, 
  target_table, 
  colnames(benchmark_table_sample))
 
# evaluate networks of 2 benchmarks 
network_list <- gen_network_names(unique(benchmarks_results$network), 2)
evaluate_targets_using_benchmarks(benchmarks_results, network_list)

# evaluate the option of using all three benchmarks in a single network
evaluate_targets_using_benchmarks(benchmarks_results, "PB_0001__PB_0002__PB_0003")
}
