% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/representation.R
\name{evaluate_targets_using_clip}
\alias{evaluate_targets_using_clip}
\title{Evaluate targets by clipping.}
\usage{
evaluate_targets_using_clip(
  reserves_sf,
  reserves_id,
  representation_raster,
  target_table,
  reserve_list = c()
)
}
\arguments{
\item{reserves_sf}{sf object containing reserves to evaluate}

\item{reserves_id}{Unique id column in reserves_sf containing reserve names as strings.}

\item{representation_raster}{Raster object of the criteria layer that will be evaluated, with crs matching reserves_sf}

\item{target_table}{Data frame containing columns "class_value" and "target_km2". i.e. the output from \code{\link[=gen_targets]{gen_targets()}}. All classes in the target table are
evaluated. All class_values must match the values in the representation_raster}

\item{reserve_list}{The reserve_id values to process. Defaults to all reserve_id's in reserves_sf}
}
\value{
A tibble with columns:
\itemize{
\item{\code{class_value}: the list of class_values from the target_table}
\item{\code{area_km2}: area of the class value in the network}
\item{\code{network}: name of the network copied from network_list}
\item{\code{class_proportion}: copied from target_table}
\item{\code{target_km2}: copied from target_table}
\item{\code{prop_target_met}: area_km2/target_km2}
}
}
\description{
Similar to \code{\link[=evaluate_targets_using_catchments]{evaluate_targets_using_catchments()}} but instead of pre-calculating the areas of the criteria raster in the catchments,
this method clips the criteria raster directly to each provided polygon.
Generally used for evaluating reserves that did not come from BUILDER and do not conform to catchment boundaries.
Slower than \code{\link[=evaluate_targets_using_catchments]{evaluate_targets_using_catchments()}} for large numbers of reserves.
}
\examples{
target_table <- gen_targets(ref_poly, led_sample, 1600)
reserves <- catchments_to_benchmarks(
  benchmark_table_sample, 
  catchments_sample, 
  c("PB_0001", "PB_0002"))
evaluate_targets_using_clip(reserves, "network", led_sample, target_table)
}
