% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.R
\name{gen_network_names}
\alias{gen_network_names}
\title{Create a vector of network names.}
\usage{
gen_network_names(in_names, k, force_in = c())
}
\arguments{
\item{in_names}{Vector in input names. Usually the names of benchmarks (e.g. PB_0001) and/or user provided reserves (e.g. PA_1).}

\item{k}{Integer > 1 and <= \code{length(in_names)}. Count of names to appear in each output network name.}

\item{force_in}{Vector of names to force in to the output vector (usually a subset of in_names but could be a substring of an output
name, see examples). Only network names containing one of these vectors will be included in the output.}
}
\value{
A sf object of networks.
}
\description{
Takes a vector of benchmark or reserve names and combines them into network names using the separator: \code{"__"}.
}
\details{
Benchmark or reserve names should never include the separator: \code{"__"}, however \code{"_"} is acceptable.

By default all combinations of benchmark/reserve names will be created based on the \code{k} parameter.
The length of the output will therefore equal \code{choose(length(in_names), k)}.

Input names are ordered in the network name based using \code{sort()}.

The output vector of names can be filtered using force_in to include specific names.
}
\examples{
gen_network_names(c("PB_1", "PB_2", "PB_3", "PB_11"), 2)

gen_network_names(c("PB_1", "PB_2", "PB_3", "PB_11"), 2, "PB_1")

gen_network_names(c("PB_1", "PB_2", "PB_3", "PB_11"), 3, c("PB_1", "PB_11"))

gen_network_names(c("PB_1", "PB_2", "PB_3", "PB_11"), 3, "PB_1__PB_11")
}
