% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/representation.R
\name{gen_targets}
\alias{gen_targets}
\title{Generate target tables for representation analysis.}
\usage{
gen_targets(
  reference_sf,
  representation_raster,
  reserve_size,
  class_values = c()
)
}
\arguments{
\item{reference_sf}{sf object of the reference area we are aiming to represent}

\item{representation_raster}{Raster object of the representation layer classified into categorical classes}

\item{reserve_size}{The area in km2 that targets will sum to. Generally the approximate reserve size of reserves being evaluated.}

\item{class_values}{A vector of classes in representation_raster to generate targets for. Defaults to all classes in the representation_raster.}
}
\value{
A tibble with columns:
\itemize{
\item{class_value: the list of class_values}
\item{area_km2: the area of each class_value in the reference_sf}
\item{reserve_size: the provided reserve_size}
\item{class_proportion: area_km2/sum(area_km2)}
\item{target_km2: class_proportion * reserve_area}
}
}
\description{
Takes a reference area polygon and a classified raster for which representation will be evaluated. Calculates the proportion of each class in the reference area
and multiples by reserve size to get target in km2.
}
\examples{
gen_targets(ref_poly, led_sample, 5000)
}
