% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrology.R
\name{get_upstream_catchments}
\alias{get_upstream_catchments}
\title{Calculates all upstream catchments for each provided protected area polygon and returns as a table, with column names as the unique id
of the protected areas.}
\usage{
get_upstream_catchments(pa_sf, pa_id, catchments_sf)
}
\arguments{
\item{pa_sf}{sf object of protected area polygons.}

\item{pa_id}{String matching the unique identifier column in \code{pa_sf}.}

\item{catchments_sf}{sf object of the catchments dataset with unique identifier column: CATCHNUM .}
}
\value{
Tibble where each column name is a unique protected area id, and each row is a catchment making up the
upstream area for that protected area. Blank rows are filled with NA.
}
\description{
Calculates all upstream catchments for each provided protected area polygon and returns as a table, with column names as the unique id
of the protected areas.
}
\examples{
reserves <- catchments_to_benchmarks(
  benchmark_table_sample, 
  catchments_sample, 
  c("PB_0001", "PB_0002", "PB_0003"))
get_upstream_catchments(reserves, "network", catchments_sample)
}
