% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_benchmarks.R
\name{group_benchmarks_using_grid}
\alias{group_benchmarks_using_grid}
\title{Group benchmarks based on overlap. Groups are assigned based on polygon centroid membership within
an intersecting grid. Centroids in the same grid cell are assigned the same group id.}
\usage{
group_benchmarks_using_grid(benchmarks_sf, grid_size)
}
\arguments{
\item{benchmarks_sf}{sf object of benchmarks or reserve polygons.}

\item{grid_size}{Numeric size of grid cells to create in units matching the crs of \code{benchmarks_sf}.}
}
\value{
Vector of group ids matching the input polygons.
}
\description{
Group benchmarks based on overlap. Groups are assigned based on polygon centroid membership within
an intersecting grid. Centroids in the same grid cell are assigned the same group id.
}
\examples{
reserves <- catchments_to_benchmarks(
  benchmark_table_sample, 
  catchments_sample, 
  c("PB_0001", "PB_0002", "PB_0003"))
group_benchmarks_using_grid(reserves, 10000)
}
