% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_benchmarks.R
\name{group_networks_using_grid}
\alias{group_networks_using_grid}
\title{Group networks based on overlap of their component benchmarks. Individual benchmarks are first grouped using
\code{\link[=group_benchmarks_using_grid]{group_benchmarks_using_grid()}}. Networks are then grouped based on the combination of their component
benchmarks. Two networks with component benchmarks in matching groups will be grouped together.}
\usage{
group_networks_using_grid(network_list, benchmarks_sf, benchmark_id, grid_size)
}
\arguments{
\item{network_list}{Vector of network names to group. Must be in the format x1__x2__x3 where x1, x2 and x3 are
values in the \code{benchmarks_sf} \code{benchmark_id} column. Any number of x can be used (e.g. x1__x2,
x1__x2__x3__x4 etc.).}

\item{benchmarks_sf}{sf object of benchmarks or reserve polygons.}

\item{benchmark_id}{unique identifier column in \code{benchmarks_sf}.}

\item{grid_size}{Numeric size of grid cells to create in units matching the crs of \code{benchmarks_sf}.}
}
\value{
Vector of group ids matching \code{network_list}.
}
\description{
e.g. for two networks PA1__PA2 and PB3__PB4, if PB1 and PB3 fall in the same gride cell, and PB2 and PB4 fall
in the same grid cell, networks will receive the same grid id.
}
\examples{
benchmarks <- catchments_to_benchmarks(
  benchmark_table_sample, 
  catchments_sample, 
  colnames(benchmark_table_sample))
network_names <- gen_network_names(benchmarks$network, 2)
networks_sf <- benchmarks_to_networks(benchmarks, network_names)

group_networks_using_grid(networks_sf$network, benchmarks, "network", 10000)
}
