% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{habitat_sample}
\alias{habitat_sample}
\title{A fictional map representing habitat quality.}
\format{
A RasterLayer with values between 1 and 3.
}
\source{
\url{https://beaconsproject.ualberta.ca/}
}
\usage{
habitat_sample
}
\description{
Derived from the LED map to demonstrate various tools.
}
\keyword{datasets}
