% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{led_sample}
\alias{led_sample}
\title{A subset of the BEACONs Lake Edge Density dataset for a sample reference area.}
\format{
A RasterLayer with values between 1 and 15.
}
\source{
\url{https://beaconsproject.ualberta.ca/}
}
\usage{
led_sample
}
\description{
A classified version of the original dataset that ranged from values of 0 to 1.48
km/km2. This version groups values into 30 interval classes, but only 15 classes
exist in the sample area.
}
\details{
Lake Edge Density quantifies the linear density of terrestrial-aquatic edges and represents the abundance of
riparian habitat along large waterbodies.
LED units are measured in km/km2 within a circular moving window (e.g., 25 km 2) and includes the perimeters
of all mapped water bodies including lakes, ponds, and large streams and rivers.
}
\keyword{datasets}
