% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranking_metrics.R
\name{raster_geometric_mean}
\alias{raster_geometric_mean}
\title{Calculate geometric mean of a raster layer inside a set of reserve polygons.}
\usage{
raster_geometric_mean(reserves_sf, raster_layer)
}
\arguments{
\item{reserves_sf}{sf object of reserves in which to sum raster values.}

\item{raster_layer}{Raster object that will be clipped to the reserves, with crs matching reserves_sf.}
}
\value{
A vector of geometric means matching the order of the input \code{reserves_sf}.
}
\description{
Clips the raster layer to each reserve polygon and calculates the geometric mean (useful for non-normal distributions).
}
\examples{
reserves <- dissolve_catchments_from_table(
  catchments_sample, 
  benchmark_table_sample,
  "network")

raster_geometric_mean(reserves, led_sample)
}
