% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ref_poly}
\alias{ref_poly}
\title{Example reference area.}
\format{
A sf object with 1 feature.
}
\source{
\url{https://beaconsproject.ualberta.ca/}
}
\usage{
ref_poly
}
\description{
Polygon of an example reference area that we want to be represented.
}
\keyword{datasets}
