% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{streams_sample}
\alias{streams_sample}
\title{Example stream network.}
\format{
A sf object with 2273 features.
}
\source{
\url{https://beaconsproject.ualberta.ca/}
}
\usage{
streams_sample
}
\description{
An example Stream network file with stream segments grouped into BASINs.
}
\keyword{datasets}
