% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranking_metrics.R
\name{sum_polygon_values}
\alias{sum_polygon_values}
\title{Sum area of a polygon layer inside a set of reserve polygons.}
\usage{
sum_polygon_values(
  reserves_sf,
  reserves_id,
  sum_polygon,
  sum_polygon_groups = NULL,
  class_vals = c(),
  fill_zeros = TRUE
)
}
\arguments{
\item{reserves_sf}{sf object of reserves in which to sum raster values.}

\item{reserves_id}{String representing unique id column in \code{reserves_sf}.}

\item{sum_polygon}{sf object of one or more polygons to sum.}

\item{sum_polygon_groups}{Column name in \code{sum_polygon} holding grouping variables. If provided, area of each group is summed, otherwise
full area of \code{sum_polygon} is summed.}

\item{class_vals}{Vector of class values in \code{sum_polygon_groups} to include in the output. Defaults to all non-NA values in \code{sum_polygon_groups}.
Ignored if \code{sum_polygon_groups} is not provided.}

\item{fill_zeros}{Should missing values be included with an area of zero? Default is TRUE.}
}
\value{
A tibble listing area in km2 for \code{sum_polygon} or \code{sum_polygon_groups} in each reserve.
}
\description{
Sums the area of a polygon dataset, or of each group with the polygon dataset, within each reserve polygon.
}
\details{
Sums are returned in a tibble which can be joined to \code{reserves_sf} using a \code{dplyr::left_join()}
and \code{dplyr::pivot_wider()}.

Projections of \code{reserves_sf} and \code{sum_polygon} must match and are assumed to have units in meters.
}
\examples{
library(dplyr)
reserves <- dissolve_catchments_from_table(
  catchments_sample, 
  benchmark_table_sample,
  "network")

# Make a habitat polygon with 2 groups using catchments  
habitat <- catchments_sample \%>\%
             filter(CATCHNUM \%in\% c(191396, 191649)) \%>\%
             mutate(group = c(1,2)) \%>\%
             select(group)
             
sum_polygon_values(reserves, "network", habitat)
sum_polygon_values(reserves, "network", habitat, "group")
sum_polygon_values(reserves, "network", habitat, "group", c(1))
sum_polygon_values(reserves, "network", habitat, "group", c(2))
sum_polygon_values(reserves, "network", habitat, "group", c(2), fill_zeros = FALSE)
sum_polygon_values(reserves, "network", habitat, "group", fill_zeros = FALSE)

# To join results back to reserves
library(tidyr)
library(dplyr)
reserves \%>\%
  left_join(
    pivot_wider(
      sum_polygon_values(reserves, "network", habitat, "group"), 
      names_from = group, values_from = area_km2)
   )
}
