% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranking_metrics.R
\name{sum_raster_values}
\alias{sum_raster_values}
\title{Sum area of each raster value in a set of polygons.}
\usage{
sum_raster_values(
  reserves_sf,
  reserves_id,
  sum_raster,
  class_vals = c(),
  fill_zeros = TRUE
)
}
\arguments{
\item{reserves_sf}{sf object of reserves in which to sum raster values.}

\item{reserves_id}{String representing unique id column in \code{reserves_sf}.}

\item{sum_raster}{Raster object to sum.}

\item{class_vals}{Vector of class values in \code{sum_raster} to include in the output. Defaults to all non-NA values in \code{sum_raster}.}

\item{fill_zeros}{Should missing values be included with an area of zero? Default is TRUE.}
}
\value{
A tibble listing area in km2 for each raster value in each reserve.
}
\description{
Intersects a raster with a set of reserve polygons and sums the area of each raster value.
}
\details{
Sums are returned in a tibble which can be joined to \code{reserves_sf} using a \code{dplyr::left_join()}
and \code{dplyr::pivot_wider()}.

Projections of \code{reserves_sf} and \code{sum_raster} must match and are assumed to have units in meters.
}
\examples{
reserves <- dissolve_catchments_from_table(
  catchments_sample, 
  benchmark_table_sample,
  "network")
  
sum_raster_values(reserves, "network", habitat_sample)
sum_raster_values(reserves, "network", habitat_sample, fill_zeros = FALSE)
sum_raster_values(reserves, "network", habitat_sample, c(2,3))

# To join results back to reserves
library(tidyr)
library(dplyr)
reserves \%>\%
  left_join(
    pivot_wider(
      sum_raster_values(reserves, "network", habitat_sample), 
      names_from = value, values_from = area_km2)
   )
}
