% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vignette_benchmark_tab}
\alias{vignette_benchmark_tab}
\title{Sample benchmarks table containing 166 benchmarks.}
\format{
A data.frame object with 166 columns and 238 rows.
}
\source{
\url{https://beaconsproject.ualberta.ca/}
}
\usage{
vignette_benchmark_tab
}
\description{
A tibble containing column names of benchmark ids and rows of catchment CATCHNUM values
found in the benchmark.
}
\keyword{datasets}
