% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vignette_builder}
\alias{vignette_builder}
\title{Sample benchmarks table containing 100 benchmarks from the vignette study region.}
\format{
A data.frame object with 100 columns and 827 rows.
}
\source{
\url{https://beaconsproject.ualberta.ca/}
}
\usage{
vignette_builder
}
\description{
A tibble containing column names of benchmark ids and rows of catchment CATCHNUM values
found in the benchmark.
}
\keyword{datasets}
