% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vignette_builder_tab}
\alias{vignette_builder_tab}
\title{Sample benchmarks table containing 166 benchmarks.}
\format{
A data.frame object with 26 columns and 10 rows.
}
\source{
\url{https://beaconsproject.ualberta.ca/}
}
\usage{
vignette_builder_tab
}
\description{
A tibble containing column names of reserve ids and rows of catchment CATCHNUM values
found in the reserve.
}
\keyword{datasets}
