% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vignette_catchments}
\alias{vignette_catchments}
\title{Sample of BEACONs catchments covering FDA09EA.}
\format{
A sf object with 3263 features and 16 fields.
}
\source{
\url{https://beaconsproject.ualberta.ca/}
}
\usage{
vignette_catchments
}
\description{
A subset of the BEACONs catchments dataset intersecting the vignette reference area.
}
\keyword{datasets}
