% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vignette_intact}
\alias{vignette_intact}
\title{Intactness layer clipped to FDA09EA}
\format{
A sf object with 1 feature and 0 fields.
}
\source{
\url{https://intactforests.org/index.html}
}
\usage{
vignette_intact
}
\description{
A subset of the Intact Forest Landscpae 2020 map.
}
\keyword{datasets}
