% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vignette_led}
\alias{vignette_led}
\title{A subset of the BEACONs Lake Edge Density dataset for the vignette reference area.}
\format{
A RasterLayer with values between 1 and 24.
}
\source{
\url{https://beaconsproject.ualberta.ca/}
}
\usage{
vignette_led
}
\description{
A classified version of the original dataset that ranged from values of 0 to 1.48
km/km2. This version groups values into 30 interval classes.
}
\keyword{datasets}
