% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vignette_nalc}
\alias{vignette_nalc}
\title{North American Land Cover 2015 (30m resolution), clipped to FDA09EA.}
\format{
A RasterLayer with values between 1 and 19.
}
\source{
\url{http://www.cec.org/north-american-environmental-atlas/land-cover-30m-2015-landsat-and-rapideye/}
}
\usage{
vignette_nalc
}
\description{
13 land cover classes present in this region.
}
\keyword{datasets}
