% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vignette_reference}
\alias{vignette_reference}
\title{Vignette reference area.}
\format{
A sf object with 1 feature.
}
\source{
\url{https://beaconsproject.ualberta.ca/}
}
\usage{
vignette_reference
}
\description{
Polygon of a larger example reference area for use in vignettes.
}
\keyword{datasets}
