% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vignette_reserves}
\alias{vignette_reserves}
\title{Existing reserves clipped to FDA09EA for use in vignette.}
\format{
A sf object with 2 features and 2 fields.
}
\source{
\url{https://www.canada.ca/en/environment-climate-change/services/national-wildlife-areas/protected-conserved-areas-database.html}
}
\usage{
vignette_reserves
}
\description{
Contains three portions of Tombstone Territorial Park.
}
\keyword{datasets}
