% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vignette_streams}
\alias{vignette_streams}
\title{Sample of streams covering FDA09EA.}
\format{
A sf object with 13407 features and 1 fields.
}
\source{
\url{https://www.nrcan.gc.ca/science-and-data/science-and-research/earth-sciences/geography/topographic-information/geobase-surface-water-program-geeau/national-hydrographic-network/21361}
}
\usage{
vignette_streams
}
\description{
A subset of the National Hydro Network streams layers.
}
\keyword{datasets}
