% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{benchmarks_to_networks}
\alias{benchmarks_to_networks}
\title{Combine a set of benchmark polygons into a network.}
\usage{
benchmarks_to_networks(benchmark_table, catchments_sf, network_list)
}
\value{

}
\description{
Takes a set of individual benchmark polygons and combines them into network polygons based on a list of network names.
}
