% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{list_overlapping_benchmarks}
\alias{list_overlapping_benchmarks}
\title{List all pairs of touching/overlapping benchmarks.}
\usage{
list_overlapping_benchmarks(benchmarks_sf)
}
\arguments{
\item{benchmarks_sf}{sf object with unique id column named \code{network}, typically the output from
\code{\link[=dissolve_catchments_from_table]{dissolve_catchments_from_table()}}, possibly with additional reserves added using \code{\link[=append_reserve]{append_reserve()}}.}
}
\value{
Vector of network names constructed using the \code{"__"} separator.
}
\description{
For all touching/overlapping polygons in the input file, returns the network_name of the overlapping
pairs.
}
\details{
This can be converted to a list of pairs of individual names using \code{sep_network_names(list_overlapping_benchmarks())}.
}
\examples{
benchmarks <- dissolve_catchments_from_table(
  catchments_sample, 
  benchmark_table_sample, 
  "network",
  dissolve_list = colnames(benchmark_table_sample))
list_overlapping_benchmarks(benchmarks)

sep_network_names(list_overlapping_benchmarks(benchmarks))
}
