% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vignette_catchments}
\alias{vignette_catchments}
\title{Sample of catchments for the vignette reference area.}
\format{
A sf object with 18123 features and 1 field.
}
\source{
\url{https://beaconsproject.ualberta.ca/}
}
\usage{
vignette_catchments
}
\description{
A subset of the BEACONs catchments dataset intersecting the vignette reference area.
}
\keyword{datasets}
