% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vignette_led}
\alias{vignette_led}
\title{BEACONs Lake Edge Density dataset clipped to FDA09EA.}
\format{
A RasterLayer with values between 0 and 4.19.
}
\source{
\url{https://www.sciencebase.gov/catalog/item/56d67611e4b015c306f2710a}
}
\usage{
vignette_led
}
\description{
Lake Edge Density quantifies the linear density of terrestrial-aquatic edges and represents the abundance of
riparian habitat along large waterbodies.
LED units are measured in km/km2 within a circular moving window (e.g., 25 km2) and includes the perimeters
of all mapped water bodies including lakes, ponds, and large streams and rivers.
}
\keyword{datasets}
